+++ CHANGES.md
@@ -1,6 +1,17 @@
 This file lists major or notable changes to OpenPnP in chronological order. This is not
 a complete change list, only those that may directly interest or affect users.
 
+# 2019-06-13
+
+## New Scripting Events
+
+Added several new scripting events for pick and place events:
+
+* Nozzle.BeforePick
+* Nozzle.AfterPick
+* Nozzle.BeforePlace
+* Nozzle.AfterPlace
+
 # 2019-06-12
 
 ## Switcher Camera and Camera Interface Changes